Fix pointer truncation, endianness and allocation bugs across Krazoa Palace
Port-layer fixes for the 4->8 pointer growth and big-endian file data.
MEM1 lands above 4GB, so any pointer routed through a 32-bit int is a
guaranteed segfault once reached.

Pointer truncation
- Clear all compiler-detectable truncation in Krazoa Palace code (79 -> 0)
  and Tricky (116 -> 1, the remainder inert and documented).
- Widen GameObject.userData1/userData2 to intptr_t; several classes stash a
  GameObject* there and read it back as a pointer.
- Convert vtable calls that read the vtable pointer as int and assumed
  4-byte entries to indexed form.
- Drop object_render_legacy.h, whose macro recast the render entry point to
  take an int and truncated every object pointer.

Endianness
- Route object-specific placement reads (offset >= 0x18) through
  ObjAnim_ReadPlacement*; only the 24-byte base header is swapped at load.
- Swap RomCurveDef requiredBit/forbiddenBit, which made every gated curve
  evaluate as never passable.
- Compose DFSH shrine target ids with explicit shifts.
- Store the ApplyBlendChannels sentinel pre-swapped. It is a native local
  read through a big-endian helper, so 1121 decoded as 260 and every vertex
  past 260 decoded stack memory as morph deltas.

Allocation sizes
- Size gMapTextureOverrides and gMapBlocks from sizeof rather than retail
  byte counts; both under-allocated and corrupted adjacent heap.
- Tie getExtraSize to sizeof where the state struct grew.

Player state
- Read queuedBits and playerStatus via named fields instead of retail
  offsets 0x8b9/0x35c, which drifted to 0x97d/0x3b0.

Also records the underlying patterns in CLAUDE.md.

+++ game/include/game/objects/object.h
@@ -73,7 +73,7 @@ struct GameObject {
     u8 sphereMapIntensity; /* obj+0xF1: r=g=b konst gray of the sphere-map TEV stage (addSphereMapTexStage and the objprint fuzz/render stages); inherited parent->child; player.c ramps it 4/frame toward skyGetSlotBlendAlpha(2) as its floor */
     u8 lightColorSlot; /* obj+0xF2: sky-light / ambient object-color slot */
     u8 unkF3;
-    s32 userData1; /* obj+0xF4/0xF8: two generic per-instance scratch words. No
+    intptr_t userData1; /* obj+0xF4/0xF8: two generic per-instance scratch words. No
         engine file reads or writes them - every access is in an object-class
         file, and each class picks its own role and width: countdown timer
         (iceball/kaldachomspit/mmshwaterspike, -= timeDelta), one-shot latch
@@ -84,7 +84,7 @@ struct GameObject {
         (mmshwaterspike). Declared s32 = the widest common access; classes
         needing another type launder through a cast. Deliberately NOT given a
         role name - the role belongs to the class, not the engine. */
-    s32 userData2;
+    intptr_t userData2;
     f32 externalVelX; /* obj+0xFC..0x104: velocity imparted externally
         (carrier object's velocity / move-data velocity), added to
         anim.velocity in the localPos integration */
