+++ game/src/dlls/modgfx/114/114.c
@@ -50,7 +50,7 @@ u32 gDll72EffectResourceData[sizeof(Dll72EffectResourceView) / sizeof(u32)] = {
 
 void dll_72_spawnEffect(GameObject* sourceObj, int variant, PartFxSpawnParams* spawnParams, u32 spawnFlags) {
     ModgfxSpawnPacket packet;
-    u8* resourceData = (u8*)(int)gDll72EffectResourceData;
+    u8* resourceData = (u8*)gDll72EffectResourceData;
     GfxCmd* commands = packet.entries;
     GameObject* context;
 
@@ -163,7 +163,7 @@ void dll_72_spawnEffect(GameObject* sourceObj, int variant, PartFxSpawnParams* s
         }
     }
     (*gModgfxInterface)
-        ->spawnEffect(&packet, 0, 0x15, (u8*)(int)gDll72EffectResourceData, 0x18,
+        ->spawnEffect(&packet, 0, 0x15, (u8*)gDll72EffectResourceData, 0x18,
                       &resourceData[offsetof(Dll72EffectResourceView, triangles)], 0x154, 0);
 }
